extract Action
