+++ src/main/java/com/deloitte/elrr/services/model/Action.java
@@ -1,0 +1,9 @@
+package com.deloitte.elrr.services.model;
+
+/**
+ * Represents the possible actions that can be performed on resources.
+ * This enum defines the standard set of operations available in the system.
+ */
+public enum Action {
+    CREATE, READ, UPDATE, DELETE, ASSOCIATE, DISASSOCIATE
+}
